+++ Blip/Blip.csproj
@@ -6,4 +6,8 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Folder Include="src\Diagramming\" />
+    </ItemGroup>
+
 </Project>
