+++ Blip/src/Format/IStringFormatter.cs
@@ -1,4 +1,4 @@
-namespace Blip.Formatters;
+namespace Blip.Format;
 
 public interface IStringFormatter {
     public char[] FormatString(string str, int width, int height);
