+++ Blip/src/Format/TruncationFormatter.cs
@@ -1,4 +1,4 @@
-namespace Blip.Formatters;
+namespace Blip.Format;
 
 public class TruncationFormatter(Alignment alignment) : IStringFormatter {
     public char[] FormatString(string str, int width, int height) {
