+++ Blip/src/Transform/OriginOnlyTransform.cs
@@ -1,4 +1,4 @@
-namespace Blip.Transforms;
+namespace Blip.Transform;
 
 public class OriginOnlyTransform : IDrawTransform {
     public char Transform(char sourceChar, char destChar) {
